requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dependency-injector==4.48.1
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-eventstream==5.3.2
django-fernet-fields==0.6
django-grip==3.5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gripcontrol==4.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pubcontrol==3.5.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rq==2.4.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
